Working windows build compiling with Visual Studio 2010. I ditched MinGW because I couldn't get the linker to link with OpenGL properly. The included VS2010 solution will compile the game.
Because the VS solution and Makefile are not in sync, I would not suggest anyone use it for active development. It will be much easier to treat the windows build as a second class citizen and backport any changes from the Makefile than trying to keep them in sync.

THIS MAY BREAK THE LINUX BUILD! I have not tested this on my Linux machine yet... but I will as soon as this is committed and I sync it over to my laptop.

If you are including glut from now on, please take a look at one of the graphics files to see how to do it. You need to include that whole #ifdef _WIN32 through the #endif.

There is something wonky with the physics in the windows build. Objects seem to glide for a while before coming to rest. I'm going to verify this does *not* match the linux build before opening a bug.